move core plugin code to its own file
